pyproject.toml
[project]
name = "fitness-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=5.5",
    "flask>=3.1.1",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]
